moved replacer to it's own file
